packages/examples/twitter-xai/python/requirements.txt
# elizaOS X Agent Requirements

# Core elizaOS
elizaos>=2.0.0a4

# Plugins
elizaos-plugin-xai>=2.0.0a4
elizaos-plugin-sql>=2.0.0a4

# Environment variables
python-dotenv>=1.0.0

# Testing (optional but consistent with other examples)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

